system: rewrite the default route handling
With this, there is consistency in setting the default route
and associated marker file.  Nothing is done before the decision
is final, so we may not end up with a missing route after an
attempt to reconfigure for something that doesn't work anyway.

DHCPv4 and PPP still do this manually, but we will have to route
them through the same logic instead of them manually handling their
routes.  6rd and 6to4 may have a similar issue as well.

While here, swap the system_routing_configure() arguments to
align with the way other configure functions present their
arguments (verbose print flag first).

Gateway monitoring and default switching improvement since 2016.
Push $keep into default handling, it may be useful in general to
avoid spurious switches.  It could also be a default probably.