work in progress pluggable backup providers for https://github.com/opnsense/core/pull/2251
Define a basic interface for backup providers and try to move our GoogleDrive backup in the process.
The basic idea is there, now we need to attach this to the current user interface (don't want to rework the legacy backup/restore page now) and move the last pieces for the google backup feature over to this.
Next we should be able to iterate over our providers and request a backup for the ones that are activated.

The responisibilty for what to backup lies within the provider at the moment, to keep things compatible and simple for the caller.