system: move factory default to configuration
Ideally, we should hide this under a different page and pop up
a confirmation dialog instead of presenting the dialog right away.

We also need a "reboot dialog" that never ends and shut down instead
of reboot.  While here, it would be good to share this dialog with
the shutdown page.