MVC, obey 'user-config-readonly' for subclasses of ApiMutableModelControllerBase.
Eventually (19.7?), user-config-readonly should go to the features of past times, it has has quite some downsides and will keep haunting everybody as long as it's there.

This patch makes sure ApiMutableModelControllerBase->save() adds an additional ACL check before doing the actual save, raising a user exception when not allowed.
Not all our classes use this, but can probably do so, same counts for plugins. We are not, under any circumstances, are going to move an acl check to the Config or model classes, since these have no relationship with the user.