dpinger: only monitor for existing entries of enabled interfaces
I have a little beef with the new code: the System: Gateways: Single
page listing should always list existing config.xml gateways, but it
does not.  So it looks like a gateway was deleted when in fact it is
being hidden.  This is definitely a visibility departure from the
previous code.