radvd: remove rasend/ranosend checkbox; closes #3365
I'm removing this after more research in the hopes that nobody
will be complaining about it.  The sad side is that this is a
simple patch and a PR would have been an effective way to review
and decide on removal.  This way, core team members are forced
to spend time otherwise spent on reasearching impact *and* also
writing the code to do it which from a divide-and-conquer view
makes this code change worse than it could have been.  ;)