interfaces: remove the "Directly send SOLICIT" option
We already do a more-or-less hybrid approach by starting rtsold
even if it isn't used at all.  Now we also have ISPs which do
not seem to send router advertisements after successful connect
so that the reconnect misses the HUP for dhcp6c to fix the
connectivity again.

To change that remove the option and its only conditional to
behave unconditionally which has few reasons to cause regressions.