Firewall / Rules - fix parser to support "maximum" ranges. Currently it's possible to input values like "80-any", in which case the lower bound is ignored (only "any" sticks). Since it's technically not very problematic to support ranges where "any" means either upper or lower bound, we best make sure the values inserted are at least used. (as soon as it's a range, it can't be an alias anyway due to the colon)
There is a small risk of regressions in existing setups where people accidentally used the unsupported input, but this looks like a reasonable risk.

The other option would be to extend the validation and reject the input.

This commit doesn't handle the user input for the "any - static port" type, which currently doesn't seem to be supported. As discussed with @kulikov-a, who might have ideas about the user input part when he has some time.