interfaces: stop mdp5 before starting
An ancinet dragon was woken from its slumber while reworking the interface
configuration code.  All things considered this is more than reasonable
although we do not yet know how this condition can be reached now as opposed
to 21.7 and if the inline termination of mdp5 will not invoke any sort of
deconfiguration (ppp-linkdown) that would harm the impending start of the
service.

PR: https://forum.opnsense.org/index.php?topic=26652.0