system: simplify previous a bit
We can set gateway to null as it is ignored when setting host
route.  Also adhere to logic for interface selection in both
DNS modes, but only exclude from configuration list for dynamic
ones.

For the time being we don't need static/dymanic annotation
as the system doesn't care by design and we don't filter it
anywhere else.  Might be something for the status page, but
that's a larger change to level expectations vs. dyamic
connectivity properties.

A couple of style issues here and there as well.