Firewall Log Widget Header Tweak (#5462)
Reduce line wrap:
Remove "Dest " from the destination port header.  Provides little value at the expense of increase line wrap.  In this context and column position (to right of destination address) it is obvious to be the destination port.  Source port in this context and column position would make no sense.