interfaces: remove out-of-band _defaultgw file writes; closes #5900
IPv4 doesn't do this indicating that IPv6 shouldn't as well.
Also, ifctl is not handling this file either so we might as
well let system.inc deal with this.

PPPoE and DHCP still read the file but wouldn't consider this
a huge problem.

Also now makes sure the scope is added to link-local gateways,
which was a problem previously reported by @maurice-w.