interfaces: fix rename after proto addition
'proto' was set on to be deleted VLANs preventing them to be
deleted and also causing the new VLAN not to gain the proper
settings.  Likely needs a reboot on systems where this bad
state was reached since the system will not delete stray VLAN
devices it does not know about anymore/failed to delete before.

PR: https://forum.opnsense.org/index.php?topic=32225.0