system: prevent spurious far gateway detect
Reported for DHCP in #5493 consider that all connectivity modes
that create a router file to be eligable for this only.  VIPs
do overcomplicate this but that is also where the gateway has
to be explicit instead of being provided at a later time, which
makes the validations fail which is why we end up here.

Also move the current route check to avoid some work.  The check
is not 100% when flipping the far behaviour but doing less work
and keeping the current behaviour is fine.