Revert "unbound: fixup defaults in migration and namespace changes"
This reverts commit e61123ec46f85424f14c57ebb63412d87bc48500 to be able
to test the migration easily.  We didn't set the full data (at least
local_zone_type was missing) but then again we do not have to.

Also seeing that we did not have OPNsense section before or version
attribute let us try to keep this status quo for a little longer.