VPN: IPsec: Mobile Clients - move Xauth out of "IKE Extensions" block. closes https://github.com/opnsense/core/issues/6481
Since "strongswan.conf" applies to both type of tunnels, make sure we can configure some shared settings for both options here (tunnels/connections). Eventually more settings might move out of the "IKE Extension" block, but for now it should be enough to isolate Xauth. The impact of configuring xauth when not being used is likely small, so when connections are used we always provide xauth-pam settings (the connection determines if it's actually used).

Removed the "is mobile" enabled in Auth/Services/IPsec.php, when only legacy is used, the behaviour should be the same (as xauth-pam isn't configured).