system: finish simplifying plugins_run()
We don't need to mask parameters here.  The providers should
use the same argument count or default parameters if needed.
A single call cannot pass different arguments for different
providers of the same run task.

For most cases a single provider is set anyway or the data is
simple enough.

The reason this code existed was because plugins_run() was once
copied from plugins_configure() which may require this behaviour
but also has no return data.