Firewall: Settings: Advanced - add "set debug" and "set keepcounters" options in the advanced section.
in some cases it can be practical to retrieve more debug output from the packet filter, in which case one can now enable this in the gui. The keepcounters toggle is added recently, it looks like an interesting option to being able to keep the rule statistics over reloads, if someone would like to try this, it makes sense to expose it to the gui as well (off by default).