system: refactor far gateway handling out of default route handling
We need far gateway routes (interface routes) for each gateway not
directly attached to the network or else the monitoring for it fails
which prevents a default route from being added which would fix it.

Since system_default_route() is private nowadays allow to pass down
the gateway struct which simplifies/speeds up the process.

We also pull in "configctl -- interface routes list -n json" data when
doing a reconfiguration to check if the interface route needs to be
touched and we can also use this check for default route and perhaps even
static routes later on.