configd - support streaming output using the "stream_output" handler.
This commit adds the connection to the action.execute() method which offers the ability to flush data back to the client instantly without having to wait for the call to end. As the existing types are only slighly modified, the impact of this change should be rather low. To demonstrate the streaming capability, also adjust configctl to flush output when received.

Example usage:

[find]
command:find / ; xxx
parameters:
type:stream_output
message:test stream output